!165 Fix compile warnings

+++ engine/pipeline/filters/common/dump_buffer.cpp
@@ -90,7 +90,7 @@ void DumpBufferToLog(const char* desc, const std::shared_ptr<Plugin::Buffer>& bu
     char* dstPtr = tmpStr;
     int len;
     const uint8_t* p = buffer->GetMemory()->GetReadOnlyData();
-    for (int i = 0; i < realDumpSize; i++) {
+    for (size_t i = 0; i < realDumpSize; i++) {
         len = snprintf_s(dstPtr, 3, 2, "%02x", p[i]); // max write 3 bytes, string len 2
         FALSE_RET_MSG(len > 0 && len <= 2, "snprintf_s returned unexpected value " PUBLIC_LOG_D32, len); // max len 2
         dstPtr += len;
